requirements.txt
streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-google-community>=0.0.10
langchain-ollama>=0.0.1  
openai>=1.0.0  